requirements.txt
azure-functions
openai>=1.40.0
azure-cosmos>=4.9.0
